+++ ipaccount.cpp
@@ -317,7 +317,7 @@ inline void Ipacc::add_octets(time_t timestamp, unsigned int saddr, unsigned int
 	key = buf;
 
 	if(last_flush_interval_time != cur_interval_time &&
-	   (time(NULL) - cur_interval_time) > opt_ipacc_interval / 5) {
+	   (time(NULL) - cur_interval_time) > (unsigned int) opt_ipacc_interval / 5) {
 		int saveIndexIpaccBuffer = indexIpaccBuffer == 0 ? 1 : 0;
 		if(!ATOMIC_FETCH_AND_ADD(&sync_save_ipacc_buffer[saveIndexIpaccBuffer], 1)) {
 			last_flush_interval_time = cur_interval_time;
